+++ src/components/Cards/index.module.css
@@ -1,0 +1,18 @@
+.Container-Menus{
+    display: flex;
+  width: 100%;
+  height: 100vh;
+  overflow-x: scroll;
+  scroll-snap-type: x mandatory;
+  gap: 20px;
+  margin-top: 20vh;
+}
+
+.Menus{
+    flex: 0 0 80%;
+    width: 70vw;
+    height: 70vh;
+    object-fit: cover;
+    background-color: rgb(25, 72, 72);
+    scroll-snap-align: center;
+}
